Refactor Android impression count Integer to Double
This better aligns with what React Native expects and what Gson produces
when stringifying JSON. Previously, there would be mismatch type errors
between Double and Integer.

+++ packages/react-native-bridge/android/react-native-bridge/src/main/java/org/wordpress/mobile/ReactNativeGutenbergBridge/GutenbergBridgeJS2Parent.java
@@ -169,6 +169,6 @@ public interface GutenbergBridgeJS2Parent extends RequestExecutor {
 
     void requestBlockTypeImpressions(ReadableArray newBlockTypes, BlockTypeImpressionsCallback blockTypeImpressionsCallback);
 
-    void setBlockTypeImpressionCount(String name, Integer count);
+    void setBlockTypeImpressionCount(String name, Double count);
 
 }
